Refacto lodTree code architecture
Grouping methods are now in the group.py file
Polygon extrusion is now in the polygon_extrusion.py file
Loa and Lod1 inherit from LodNode to create nodes with specific geometries
