+++ py3dtilers/Common/tileset_creation.py
@@ -1,7 +1,7 @@
 import numpy as np
 from py3dtiles import B3dm, BatchTable, BoundingVolumeBox, GlTF
 from py3dtiles import Tile, TileSet
-from ..Common import create_lod_tree
+from ..Common import LodTree
 from ..Common import createTextureAtlas
 
 
@@ -9,7 +9,7 @@ def create_tileset(objects_to_tile, also_create_lod1=False, also_create_loa=Fals
     """
     Recursively creates a tileset from the nodes of a LodTree
     """
-    lod_tree = create_lod_tree(objects_to_tile, also_create_lod1, also_create_loa, loa_path, with_texture)
+    lod_tree = LodTree.create_lod_tree(objects_to_tile, also_create_lod1, also_create_loa, loa_path, with_texture)
     tileset = TileSet()
     centroid = lod_tree.centroid
     for root_node in lod_tree.root_nodes:
